existwithEmail, existWithId unit test

+++ src/main/java/com/bachar/customer/CustomerController.java
@@ -26,7 +26,7 @@ public class CustomerController {
 
     @GetMapping("{customerId}")
     public Customer getCustomer(
-            @PathVariable("customerId") Integer customerId) {
+            @PathVariable("customerId") Long customerId) {
         return customerService.getCustomer(customerId);
     }
 
@@ -37,13 +37,13 @@ public class CustomerController {
 
     @DeleteMapping("{customerId}")
     public void deleteCustomer(
-            @PathVariable("customerId") Integer customerId) {
+            @PathVariable("customerId") Long customerId) {
         customerService.deleteCustomer(customerId);
     }
 
     @PutMapping("{customerId}")
     public void updateCustomer(
-            @PathVariable("customerId") Integer customerId, @RequestBody CustomerUpdateRequest updateRequest) {
+            @PathVariable("customerId") Long customerId, @RequestBody CustomerUpdateRequest updateRequest) {
         customerService.updateCustomer(customerId, updateRequest);
     }
 }
