+++ src/main/java/com/bachar/customer/CustomerService.java
@@ -22,7 +22,7 @@ public class CustomerService {
         return customerDao.selectAllCustomers();
     }
 
-    public Customer getCustomer(Integer customerId) {
+    public Customer getCustomer(Long customerId) {
         return customerDao.selectCustomerById(customerId).orElseThrow(() -> new ResourceNotFoundException("Customer with id %d is not found".formatted(customerId)));
     }
 
@@ -38,14 +38,14 @@ public class CustomerService {
         customerDao.insertCustomer(customer);
     }
 
-    public void deleteCustomer(Integer customerId) {
+    public void deleteCustomer(Long customerId) {
         if (!customerDao.existPersonWithId(customerId)) {
             throw new ResourceNotFoundException("Customer with id %d is not found".formatted(customerId));
         }
         customerDao.deleteCustomer(customerId);
     }
 
-    public void updateCustomer(Integer customerId, CustomerUpdateRequest updateRequest) {
+    public void updateCustomer(Long customerId, CustomerUpdateRequest updateRequest) {
         Customer customer = getCustomer(customerId);
         boolean changes = false;
